add filesize function to get size of dols.
change homebrew icon image and move it a bit.

hide the homebrew button if godmode!=1

maybe fix issue 722.  i don't have a spare drive to test it.

fix grammatical error in the settings.

fix long game name overflow on the gameinfo screen

switch the homebrew buttons to use the new HBCwindowprompt().  none of the meta.xml data is passed yet so the prompt looks pretty naked.